+++ packages/in_app_purchase/in_app_purchase_android/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.3.2
 
 * Adds UserChoiceBillingapis to platform addition.
+* Updates minimum supported SDK version to Flutter 3.13/Dart 3.1.
 
 ## 0.3.1